backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.27
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
celery==5.3.6
redis==5.0.1
pillow==10.2.0
imagehash==4.3.1
numpy==1.26.4
tenacity==8.2.3
python-multipart==0.0.7
alembic==1.13.0
openai==1.13.3
httpx==0.27.0
insightface==0.7.3
mediapipe==0.10.14
opencv-python==4.10.0.84
